pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gecko-server"
version = "0.1.0"
description = "Gecko: OpenAPI-driven mock tool server"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Zeyu Zhang" }
]
dependencies = [
  "camel-ai[all]==0.2.49",
  "fastapi==0.115.12",
  "uvicorn==0.34.0",
  "starlette>=0.45.0",
  "pydantic>=2.9.0",
  "requests",
  "python-dotenv>=1.0.0",
  "openapi-core>=0.19.0",
  "openapi_spec_validator==0.7.1",
  "PyYAML==6.0.2",
  "typing-extensions>=4.12.0",
  "tiktoken==0.5.2"
]

[project.scripts]
gecko-server = "run_gecko_server:main"

[tool.setuptools]
py-modules = ["run_gecko_server"]
packages = ["gecko", "gecko.core", "gecko.handlers", "gecko.middleware", "gecko.schemas", "gecko.tools", "gecko.utils", "utils"]
